fixture and mock refactor for transport tests
