+++ tests/mocks/__init__.py
@@ -1,6 +1,5 @@
 """Mock implementations for testing."""
 
-from .fixtures import mock_session_path, mock_telethon
-from .transports import create_mock_telethon
+from .transports.telethon import mock_session_path, mock_telethon, create_mock_telethon
 
 __all__ = ['mock_session_path', 'mock_telethon', 'create_mock_telethon']
